Removed unnecessary files. Combined files that didn't need to be broken up in the first place.
