+++ client/source/scripts/components/prompt.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import Paper from 'material-ui/Paper';
+import RaisedButton from 'material-ui/RaisedButton';
+import { Link } from 'react-router';
 import { grey600 } from 'material-ui/styles/colors';
-import QuestionGender from './QuestionGender.jsx';
-import QuestionAge from './QuestionAge.jsx';
-import QuestionGoal from './QuestionGoal.jsx';
 
 const stylePaper = {
   height: '80%',
@@ -14,6 +13,10 @@ const stylePaper = {
   textAlign: 'center',
   display: 'block',
   color: grey600,
+};
+
+const styleButton = {
+  margin: 12,
 };
 
 const answer = {
@@ -31,13 +34,21 @@ class PromptBox extends React.Component {
 
   render() {
     return (
-      <div>
         <Paper className="prompt-background" style={stylePaper} zDepth={3}>
-          <div>
-            <QuestionGoal answer={this.state.answer} />
-          </div>
-        </Paper>
+      <div>
+        <h2>What is your goal?</h2>
+        <Link to="/workout1">
+          <RaisedButton
+            label="Build Muscle" primary style={styleButton}
+          />
+        </Link>
+        <Link to="/workout2">
+          <RaisedButton
+            label="Lose Fat" secondary style={styleButton}
+          />
+        </Link>
       </div>
+        </Paper>
     );
   }
 }
